add cocos2dx v4 codes

+++ frameworks/cocos2d-x/extensions/GUI/CCControlExtension/CCControlSaturationBrightnessPicker.cpp
@@ -32,7 +32,6 @@
  */
 
 #include "CCControlSaturationBrightnessPicker.h"
-#include <cmath>
 
 NS_CC_EXT_BEGIN
 
@@ -105,14 +104,14 @@ void ControlSaturationBrightnessPicker::updateWithHSV(HSV hsv)
     hsvTemp.v = 1;
     
     RGBA rgb = ControlUtils::RGBfromHSV(hsvTemp);
-    _background->setColor(Color3B((GLubyte)(rgb.r * 255.0f), (GLubyte)(rgb.g * 255.0f), (GLubyte)(rgb.b * 255.0f)));
+    _background->setColor(Color3B((uint8_t)(rgb.r * 255.0f), (uint8_t)(rgb.g * 255.0f), (uint8_t)(rgb.b * 255.0f)));
 }
 
 void ControlSaturationBrightnessPicker::updateDraggerWithHSV(HSV hsv)
 {
     // Set the position of the slider to the correct saturation and brightness
-    Vec2 pos(_startPos.x + boxPos + (boxSize*(1 - hsv.s)),
-                              _startPos.y + boxPos + (boxSize*hsv.v));
+    Vec2 pos(_startPos.x + boxPos + (boxSize*(1 - (float)hsv.s)),
+                              _startPos.y + boxPos + (boxSize*(float)hsv.v));
     
     // update
     updateSliderPosition(pos);
@@ -155,8 +154,8 @@ void ControlSaturationBrightnessPicker::updateSliderPosition(Vec2 sliderPosition
     else if (sliderPosition.y > _startPos.y + boxPos + boxSize)        sliderPosition.y = _startPos.y + boxPos + boxSize;
     
     // Use the position / slider width to determine the percentage the dragger is at
-    _saturation = 1.0f - std::fabs((_startPos.x + (float)boxPos - sliderPosition.x)/(float)boxSize);
-    _brightness = std::fabs((_startPos.y + (float)boxPos - sliderPosition.y)/(float)boxSize);
+    _saturation = 1.0f - fabs((_startPos.x + (float)boxPos - sliderPosition.x)/(float)boxSize);
+    _brightness = fabs((_startPos.y + (float)boxPos - sliderPosition.y)/(float)boxSize);
 }
 
 bool ControlSaturationBrightnessPicker::checkSliderPosition(Vec2 location)
